// src/constants/enums/createPost.ts
export enum CreatePostForm {
  id = "create_post_id",

  autocomplete = "off",

  title_input_id = "title_input_id",
  title_input_name = "title_input",
  title_input_label = "",
  title_input_role = "title-input",
  title_input_arial_label = "title-input",
  title_input_placeholder = "I'm happy",
  title_input_type = "text",

  content_input_id = "content_input_id",
  content_input_name = "content_input",
  content_input_label = "",
  content_input_role = "content-input",
  content_input_arial_label = "content-input",
  content_input_placeholder = "What's on your mind, %user_firsname",
  content_input_type = "text",

  post_images_input_file_id = "post_images_input_id",
  post_images_input_file_name = "post_images_input_name",
  post_images_input_file_arial_label = "Upload posts picture",
  post_images_input_file_type = "file",
}